+++ Source/Krypton Toolkit/Krypton.Toolkit.Suite.Extended.Error.Reporting/Classes/Core/FileSaveResult.cs
@@ -1,0 +1,7 @@
+﻿namespace Krypton.Toolkit.Suite.Extended.Error.Reporting;
+
+internal class FileSaveResult
+{
+    public bool Saved { get; set; }
+    public Exception Exception { get; set; }
+}