+++ Source/Krypton Toolkit/Krypton.Toolkit.Suite.Extended.Error.Reporting/Classes/Templates/AssemblyRef.cs
@@ -1,0 +1,7 @@
+﻿namespace Krypton.Toolkit.Suite.Extended.Error.Reporting;
+
+public class AssemblyRef
+{
+    public string Name { get; set; }
+    public string Version { get; set; }
+}